requirements.txt
torch
torchvision
Pillow
ftfy
regex
tqdm

